08/02/2020 : Corrections de bugs

+++ A2LParser/src/a2l/RecordLayout.java
@@ -144,7 +144,7 @@ public final class RecordLayout implements A2lObject, Comparable<RecordLayout> {
         private final AdressType adressType;
 
         public FncValues(List<String> parameters) {
-            this.position = (byte) Integer.parseInt(parameters.get(0));
+            this.position = Byte.parseByte(parameters.get(0));
             this.dataType = DataType.getDataType(parameters.get(1));
             this.indexMode = IndexMode.getIndexMode(parameters.get(2));
             this.adressType = AdressType.getAdressType(parameters.get(3));
@@ -170,7 +170,7 @@ public final class RecordLayout implements A2lObject, Comparable<RecordLayout> {
         private final AdressType adressType;
 
         public AxisPtsX(List<String> parameters) {
-            this.position = (byte) Integer.parseInt(parameters.get(0));
+            this.position = Byte.parseByte(parameters.get(0));
             this.dataType = DataType.getDataType(parameters.get(1));
             this.indexOrder = IndexOrder.getIndexOrder(parameters.get(2));
             this.adressType = AdressType.getAdressType(parameters.get(3));
@@ -194,7 +194,7 @@ public final class RecordLayout implements A2lObject, Comparable<RecordLayout> {
         private final DataType dataType;
 
         public NoAxisPtsX(List<String> parameters) {
-            this.position = (byte) Integer.parseInt(parameters.get(0));
+            this.position = Byte.parseByte(parameters.get(0));
             this.dataType = DataType.getDataType(parameters.get(1));
         }
 
@@ -211,7 +211,7 @@ public final class RecordLayout implements A2lObject, Comparable<RecordLayout> {
         private final AdressType adressType;
 
         public AxisPtsY(List<String> parameters) {
-            this.position = (byte) Integer.parseInt(parameters.get(0));
+            this.position = Byte.parseByte(parameters.get(0));
             this.dataType = DataType.getDataType(parameters.get(1));
             this.indexOrder = IndexOrder.getIndexOrder(parameters.get(2));
             this.adressType = AdressType.getAdressType(parameters.get(3));
@@ -235,7 +235,7 @@ public final class RecordLayout implements A2lObject, Comparable<RecordLayout> {
         private final DataType dataType;
 
         public NoAxisPtsY(List<String> parameters) {
-            this.position = (byte) Integer.parseInt(parameters.get(0));
+            this.position = Byte.parseByte(parameters.get(0));
             this.dataType = DataType.getDataType(parameters.get(1));
         }
 
@@ -250,7 +250,7 @@ public final class RecordLayout implements A2lObject, Comparable<RecordLayout> {
         private final DataType dataType;
 
         public SrcAddrX(List<String> parameters) {
-            this.position = (byte) Integer.parseInt(parameters.get(0));
+            this.position = Byte.parseByte(parameters.get(0));
             this.dataType = DataType.getDataType(parameters.get(1));
         }
 
@@ -265,7 +265,7 @@ public final class RecordLayout implements A2lObject, Comparable<RecordLayout> {
         private final DataType dataType;
 
         public SrcAddrY(List<String> parameters) {
-            this.position = (byte) Integer.parseInt(parameters.get(0));
+            this.position = Byte.parseByte(parameters.get(0));
             this.dataType = DataType.getDataType(parameters.get(1));
         }
 
@@ -280,7 +280,7 @@ public final class RecordLayout implements A2lObject, Comparable<RecordLayout> {
         private final DataType dataType;
 
         public NoRescaleX(List<String> parameters) {
-            this.position = (byte) Integer.parseInt(parameters.get(0));
+            this.position = Byte.parseByte(parameters.get(0));
             this.dataType = DataType.getDataType(parameters.get(1));
         }
 
@@ -295,7 +295,7 @@ public final class RecordLayout implements A2lObject, Comparable<RecordLayout> {
         private final DataSize dataSize;
 
         public Reserved(List<String> parameters) {
-            this.position = (byte) Integer.parseInt(parameters.get(0));
+            this.position = Byte.parseByte(parameters.get(0));
             this.dataSize = DataSize.getDataSize(parameters.get(1));
         }
 
@@ -312,9 +312,9 @@ public final class RecordLayout implements A2lObject, Comparable<RecordLayout> {
         private final AdressType adressType;
 
         public AxisRescaleX(List<String> parameters) {
-            this.position = (byte) Integer.parseInt(parameters.get(0));
+            this.position = Byte.parseByte(parameters.get(0));
             this.dataType = DataType.getDataType(parameters.get(1));
-            this.maxNumberOfRescalePairs = (short) Integer.parseInt(parameters.get(2));
+            this.maxNumberOfRescalePairs = Short.parseShort(parameters.get(2));
             this.indexOrder = IndexOrder.getIndexOrder(parameters.get(3));
             this.adressType = AdressType.getAdressType(parameters.get(4));
         }
